backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trackforge"
version = "0.1.0"
description = "Self-hosted music request and acquisition platform"
requires-python = ">=3.12"
dependencies = [
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",

    # Redis / Queue
    "redis>=5.2.0",
    "arq>=0.26.0",

    # Config
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",

    # Auth
    "bcrypt>=4.2.0",
    "python-jose[cryptography]>=3.3.0",

    # HTTP client
    "httpx>=0.28.0",

    # Music tools
    "mutagen>=1.47.0",
    "pyacoustid>=1.3.0",

    # Utilities
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["trackforge"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
